moving MainPage.js to containers
