+++ src/containers/MainPage.js
@@ -1,9 +1,9 @@
 import React from 'react';
 import { connect } from 'react-redux';
 
-import Preview from './Preview';
-import Container from './ui/Container';
-import Column from './ui/Column';
+import Preview from '../components/Preview';
+import Container from '../components/ui/Container';
+import Column from '../components/ui/Column';
 
 function mapStateToProps(state) {
   return {
