+++ src/routes.js
@@ -1,7 +1,7 @@
 import React from 'react';
 import { IndexRoute, Route } from 'react-router';
 import App from './containers/App';
-import MainPage from './components/MainPage';
+import MainPage from './containers/MainPage';
 
 export default (
   <Route path="/" component={ App }>
